requirements.txt
asgiref==3.8.1
asttokens==3.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
decorator==5.2.1
Django==5.2
django-extensions==4.1
django-ranged-response==0.2.0
django-rosetta==0.10.2
django-simple-captcha==0.6.2
et_xmlfile==2.0.0
executing==2.2.0
idna==3.10
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
openpyxl==3.1.5
parso==0.8.4
pillow==11.2.1
polib==1.2.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pure_eval==0.2.3
Pygments==2.19.1
python-dotenv==1.1.1
requests==2.32.3
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
